ms_users/requirements.txt
pydantic>=2.7.1,<2.8.0
fastapi>=0.110.3,<0.111.0
sqlmodel>=0.0.18,<0.1.0
asyncpg>=0.29.0,<0.30.0
uvicorn>=0.29.0,<0.30.0
bcrypt>=4.1.2,<4.2.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
cryptography>=42.0.5,<42.1.0
httpx>=0.27.0,<0.28.0
alembic>=1.13.1,<1.14.0
pytest>=8.2.0,<8.3.0
pytest-mock>=3.14.0,<3.15.0
pytest-cov>=5.0.0,<5.1.0
pytest-httpx>=0.30.0,<0.31.0
pytest-asyncio>=0.23.6,<0.24.0
python-multipart>=0.0.9,<0.0.10
aiofiles>=23.2.1,<23.3.0
aio-pika>=9.4.1,<9.5.0